requirements.txt
Scrapy==0.24.4
Twisted==14.0.2
cffi==0.8.6
characteristic==14.2.0
cryptography==0.6.1
cssselect==0.9.1
gnureadline==6.3.3
ipython==2.3.1
lxml==3.4.1
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
python-dateutil==2.3
queuelib==1.2.2
selenium==2.44.0
service-identity==14.0.0
six==1.8.0
w3lib==1.10.0
wsgiref==0.1.2
zope.interface==4.1.1
